+++ SBTP/View/JCXZ/TPC_XCSJ.xaml.cs
@@ -96,7 +96,7 @@ namespace SBTP.View.JCXZ
             {
                 DataPoint point = new DataPoint();
                 point.XValue = i;
-                point.YValues = new double[] { list[i].STL };
+                point.YValues = new double[] { (double)list[i].STL };
                 point.AxisLabel = string.Format("{0}{1}{2}({3})", list[i].YCZ, list[i].XCH, list[i].XCXH, list[i].YXHD);
                 point.Label = list[i].STL.ToString();
                 series.Points.Add(point);
